move clock stuff to separate file
